organize schema options into one file
