+++ schema.go
@@ -109,46 +109,3 @@ func (s *Schema) Exec(raw Raw) {
 func (s *Schema) Do(fn Do) {
 	s.add(fn)
 }
-
-// Options options for table, column and index.
-type Options string
-
-func (o Options) applyTable(table *Table) {
-	table.Options = string(o)
-}
-
-func (o Options) applyColumn(column *Column) {
-	column.Options = string(o)
-}
-
-func (o Options) applyIndex(index *Index) {
-	index.Options = string(o)
-}
-
-func (o Options) applyKey(key *Key) {
-	key.Options = string(o)
-}
-
-// Optional option.
-// when used with create table, will create table only if it's not exists.
-// when used with drop table, will drop table only if it's exists.
-type Optional bool
-
-func (o Optional) applyTable(table *Table) {
-	table.Optional = bool(o)
-}
-
-func (o Optional) applyIndex(index *Index) {
-	index.Optional = bool(o)
-}
-
-// Raw string
-type Raw string
-
-func (r Raw) internalMigration()       {}
-func (r Raw) internalTableDefinition() {}
-
-// Do used internally for schema migration.
-type Do func(Repository) error
-
-func (d Do) internalMigration() {}
